Fix passing of 64-bit address to disasm test

+++ disassembler/gofer.c
@@ -8,7 +8,7 @@
 #include "encodings.h"
 #include "arm64dis.h"
 
-void disassemble(uint32_t addr, uint8_t *data, int len, char *result, bool verbose)
+void disassemble(uint64_t addr, uint8_t *data, int len, char *result, bool verbose)
 {
 	Instruction instr;
 	memset(&instr, 0, sizeof(instr));
